Add foreword mapping the translators and the stories they worked on
I’d been a little confused as to why the halftitle I’d copied in from the Ivan Bunin collection existed, but when I investigated further I wound the foreword I’d forgot I’d written. This commit replicates that for this collection, and describes the English translators and the stories they worked on.

Now there’s a foreword, the halftitle needs to come back in.